==> Scripts/1. Jupyter Fundamentals and Python Data Types.ipynb <==
# 01. Importing libraries 

# Import libraries
import pandas as pd
import numpy as np
import os


# Working With Python Data Types

#Question 1: Code 3 different ways of reaching a result of 100 by adding or subtracting numeric variables.

x = 50 

50+x

y = 50 

x+y

z=100

h=200

h-z

#Question 2: Code 2 floating-point variables and divide them by each other.

a=5.67
b=7.76

type(a)

a/b

#Question 3: Construct a short word made of separate strings (you’ll need to concatenate some strings!).

h= 'h'
q = 'e'
j= 'l'
k= 'o'

h+q+j+j+k

#Question 4: Construct 2 short sentences made of separate strings.

c= 'Hello'
d= 'my' 
e= 'name' 
f= 'is' 
g= 'Emily'

c+d+e+f+g

l= 'What'
m= 'your'

c+l+f+m+e
