==> Scripts/2. Data Importing & Descriptive Analysis.ipynb <==
# Contents 

1. Importing Libraries 
2. Importing Dataframe 
3. Exploring Dataframe 


# Importing Libraries 

# Import libraries 
import pandas as pd
import numpy as np 
import os

1. Turn your project folder path into a string as outlined in the Exercise text.
2. Import the “orders.csv” data set into your notebook using the os library and the vars_list shortcut, omitting the “eval_set” column in your final dataframe.

# Importing Dataframe 

df = pd.read_csv(r'C:\Users\emily\OneDrive\Documents\Instacart Basket Analysis\Data\Original Data\orders.csv', index_col = False)

path = r'C:\Users\emily\OneDrive\Documents\Instacart Basket Analysis'

path

df = pd.read_csv(os.path.join(path, 'Data', 'Original Data', 'orders.csv'), index_col = False)

# Exploring Dataframe 

vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

df = pd.read_csv(os.path.join(path, 'Data', 'Original Data', 'orders.csv'), usecols = vars_list)

vars_list

3. Print the first 5 rows of your dataframe.

df.head()

4. Import the “products.csv” file into Jupyter as df_prods (as opposed to df, which is what your “orders.csv” file is named in Jupyter). You can use the path variable you created in step 4 to import “products.csv” with the os library.

df_prods= pd.read_csv(os.path.join(path, 'Data', 'Original Data', 'products.csv'), index_col = False)

5. Print the first 20 rows of the df_prods dataframe.

df_prods.head(20)

6. Print the last 35 rows of the df_prods dataframe.

df_prods.tail(35)


7. Print the names of the columns in the df_prods dataframe.

df_prods.columns

8. Print the number of rows and columns in the df_prods dataframe.

df_prods.shape

9. What’s the maximum value of the “aisle_id” column? 
Answer: 134 

df_prods.describe()

10. What’s the data type of the “department_id” column?
Answer: int64 

df_prods.dtypes
