==> Scripts/3. Data Wrangling & Subsetting.ipynb <==
# Contents 

1. Importing Libraries 
2. Importing Dataframe 
3. Data Wrangling 
4. Data Dictionary 
5. Task 4.4 Questions 
6. Exporting Dataframe 

# Importing Libraries 

import pandas as pd
import numpy as np 
import os

# Importing Orders, Products and Departments csv

path = r'C:\Users\emily\OneDrive\Documents\Career Foundry 2022\Data Immersion\Achievment 4\Instacart Basket Analysis'

df_ords = pd.read_csv(os.path.join(path, 'Data', 'Original Data', 'orders.csv'), index_col = False)

df_prods= pd.read_csv(os.path.join(path, 'Data', 'Original Data', 'products.csv'), index_col = False)

df_dep = pd.read_csv(os.path.join(path, 'Data', 'Original Data', 'departments.csv'), index_col = False)

# Data Wrangling 

Dataframe orders.csv 

#Dropping Eval_set column from orders.csv

df_ords = df_ords.drop(columns = ['eval_set'])

df_ords

#Checking missing values (Nan)

df_ords['days_since_prior_order'].value_counts(dropna = False)

#Renaming columns for clear readability 

df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

df_ords.head()

#Making order_id column a string

df_ords['order_id'] = df_ords['order_id'].astype('str')

df_ords['order_id'].dtype

#Dataframe products.csv 
No wrangling needed 

df_prods

#Dataframe departments.csv 

df_dep

#Adjusting the column header 

df_dep.T

df_dep_t = df_dep.T

df_dep_t

df_dep_t.reset_index()

#Take the first row of df_dep_t for the header

new_header = df_dep_t.iloc[0]

new_header

df_dep_t_new = df_dep_t[1:]

df_dep_t_new

df_dep_t_new.columns = new_header

df_dep_t_new

# Data Dictionary 

data_dict = df_dep_t_new.to_dict('index')

data_dict

Exploring within Data Dictionary 

df_prods.head()

df_prods['department_id']==19

df_prods[df_prods['department_id']==19]

df_snacks =  df_prods[df_prods['department_id']==19]

df_snacks.head()

df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([19])]

df_snacks_3

# End of Data Wrangling 

# Task 4.4 Questions 

1. Find another identifier variable in the df_ords dataframe that doesn’t need to be included in your analysis as a numeric variable and change it to a suitable format.

df_ords['user_id'] = df_ords['user_id'].astype('str')

df_ords['user_id'].dtype

2. Look for a variable in your df_ords dataframe with an unintuitive name and change its name without overwriting the data frame.

df_ords.head()

df_ords.rename(columns = {'days_since_prior_order' : 'days_since_last_order_placed'}, inplace = True)

df_ords

3. Your client wants to know what the busiest hour is for placing orders. Find the frequency of the corresponding variable and share your findings. 
Answer: 10AM is the busiest hour of the day. 

df_ords['order_hour_of_day'].describe()

#Changing the variable order_hour_of_day from a float to integer 

df_ords['order_hour_of_day'] = df_ords['order_hour_of_day'].astype('int')

df_ords['order_hour_of_day'].value_counts(dropna = False)

5. Determine the meaning behind a value of 4 in the "department_id" column within the df_prods dataframe using a data dictionary.
Answer: produce 

data_dict

df_prods[df_prods['department_id']==4]

6. The sales team in your client’s organization wants to know more about breakfast item sales. Create a subset containing only the required information.

df_breakfast = df_prods[df_prods['department_id']==14]

df_breakfast

7. They’d also like to see details about customers who might be throwing dinner parties. Your task is to find all observations from the entire dataframe that include items from the following departments: alcohol, deli, beverages, and meat/seafood. You’ll need to present this subset to your client.

df_dinner_parties = df_prods.loc[df_prods['department_id'].isin([5,20,7,12])]

df_dinner_parties

8. It’s important that you keep track of total counts in your dataframes. How many rows does the last dataframe you created have?
Answer: 7650 Rows 

9. Someone from the data engineers team in Instacart thinks they’ve spotted something strange about the customer with a "user_id" of “1.” Extract all the information you can about this user.

df_ords.head()

df_ords['user_id']==1

df_ords_user_one = df_ords[df_ords['user_id']=='1']

df_ords_user_one

10. You also need to provide some details about this user’s behavior. What basic stats can you provide based on the information you have?

df_ords_user_one.describe()

# Exporting Dataframes

df_ords.to_csv(os.path.join(path, 'Data','Prepared Data', 'orders_wrangled.csv'))

df_dep_t_new.to_csv(os.path.join(path, 'Data','Prepared Data', 'departments_wrangled.csv'))
