==> Scripts/4. Data Consistency Checks.ipynb <==
# Contents 

1. Importing Libraries 
2. Importing Dataframe 
3. Consistency Checks 
4. Task 4.5 Questions 
5. Exporting Dataframe 

# Importing Libraries 

import pandas as pd
import numpy as np 
import os 

# Importing Orders and Products csv

path = r'C:\Users\emily\OneDrive\Documents\Career Foundry 2022\Data Immersion\Achievment 4\Instacart Basket Analysis'

df_prods= pd.read_csv(os.path.join(path, 'Data', 'Original Data', 'products.csv'), index_col = False)

df_ords= pd.read_csv(os.path.join(path, 'Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)

# Consistency checks with products.csv

# Finding Missing Values 

df_prods.isnull().sum()

df_nan= df_prods[df_prods['product_name'].isnull() == True]

df_nan

df_prods.shape

df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean.shape

# Searching for Duplicates 

df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

df_prods_clean.shape

df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

df_prods_clean_no_dups.to_csv(os.path.join(path, 'Data','Prepared Data', 'products_checked.csv'))

# Task 4.5 Questions 

1. Run the df.describe() function on your df_prods dataframe. Using your new knowledge about how to interpret the output of this function, share in a markdown cell whether anything about the data looks off or should be investigated further.

Answer: I think that the max values of the prices column should be investigated. It is interesting to see that there is a max price of $100,000 spent in 1 order. 

df_prods_clean_no_dups.describe()

2. Check for mixed-type data in your df_ords dataframe.

for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

3. If you find mixed-type data, fix it. The column in question should contain observations of a single data type.

Answer: Since there was no output while checking for missing data we can assume there are no mixed columns within the df_ords dataframe. 

4. Run a check for missing values in your df_ords dataframe. In a markdown cell, report your findings and propose an explanation for any missing values you find.

Answer: The missing values could be due to a customer creating an account on Instacart and only placing an order once, having no previous order history. Meaning that there will be no value within the column "days_since_last_order_placed". 

#Finding missing values

df_ords.isnull().sum()

df_ords_nan = df_ords[df_ords['days_since_last_order_placed'].isnull() == True]

df_ords_nan

5. Address the missing values using an appropriate method. In a markdown cell, explain why you used your method of choice.

Answer: It is best to leave these values as they are for now. Because they are first time orders, future orders can be placed updating the NaN values as needed. 

6. Run a check for duplicate values in your df_ords data. In a markdown cell, report your findings and propose an explanation for any duplicate values you find.

Answer: No duplicates are found in the orders_wrangled.csv dataframe. 

df_ords_dups = df_ords[df_ords.duplicated()]

df_ords_dups

7. Address the duplicates using an appropriate method. In a markdown cell, explain why you used your method of choice.

Answer: No duplicates are found in the orders_wrangled.csv dataframe. 

8. Export your final, cleaned df_prods and df_ords data as “.csv” files in your “Prepared Data” folder and give them appropriate, succinct names.

# Exporting Dataframe 

df_ords.to_csv(os.path.join(path, 'Data','Prepared Data', 'orders_wrangled_checked.csv'))

df_prods_clean_no_dups.to_csv(os.path.join(path, 'Data','Prepared Data', 'products_checked.csv'))
