==> Scripts/5. Exporting & Merging Data.ipynb <==
# Contents 

1. Importing Libraries 
2. Importing Dataframe 
3. Merging Dataframes
4. Exporting Dataframe 


# Task 4.6 Exporting and Merging Data 

Contents: 

This script contains the following points:
Exporting PKL DataFrame 'orders_products_combined.pkl'
Anwers to Task 4.6

# Importing Libraries 

import pandas as pd
import numpy as np 
import os 

1. In a new notebook, import the orders_products_combined dataframe from the pickle file you just saved.

# Importing Dataframe 

path = r'C:\Users\emily\OneDrive\Documents\Career Foundry 2022\Data Immersion\Achievment 4\Instacart Basket Analysis'

df_ords_prods_combined = pd.read_pickle(os.path.join(path, 'Data', 'Prepared Data', 'orders_products_combined.pkl'))

df_prods_checked= pd.read_csv(os.path.join(path, 'Data', 'Prepared Data', 'products_checked.csv'), index_col = False)

2. Check the shape of the imported dataframe (it should be the same as the one you exported—always check!).

Answer: the df_merged_large dataset that was exported in pkl also contains (32434489, 10) columns and rows. 

df_ords_prods_combined.shape

3. Determine a suitable way to combine the orders_products_combined dataframe with your products data set. Make sure you’re using your wrangled, cleaned, and deduped products data set stored in your “Prepared Data” folder from the previous Exercise’s task.

#Checking the output of orders_products_combined.pkl
df_ords_prods_combined.head()

df_prods_checked.head()

#Dropping unnecessary column 
df_prods_checked = df_prods_checked.drop(columns = ['Unnamed: 0'])

df_prods_checked.head()

# Merging Dataframes

#Merging products_checked.csv and orders_products_combined.pkl 

df_prods_checked

# Dropping '_merge' column from 'orders_products_combined.pkl'
df_ords_prods_combined = df_ords_prods_combined.drop(columns = ['_merge'])

df_ords_prods_combined.shape

# Merging
df_ords_prods_merged = df_ords_prods_combined.merge(df_prods_checked, on = 'product_id', indicator = True)

4. Confirm the results of the merge using the merge flag.

df_ords_prods_merged.head()

#Dropping Unnamed: 0 column as this is unnecessary 

df_ords_prods_merged = df_ords_prods_merged.drop(columns = ['Unnamed: 0'])

df_ords_prods_merged.head()

df_ords_prods_merged['_merge'].value_counts()

df_ords_prods_merged.shape

# Exporting Dataframe

5. Export the newly created dataframe as orders_products_merged in a suitable format (taking into consideration the size).

# Import to PKL
df_ords_prods_merged.to_pickle(os.path.join(path, 'Data','Prepared Data', 'orders_products_merged.pkl'))
