==> Scripts/6. Deriving New Variables.ipynb <==
# Contents 

1. Importing Libraries 
2. Importing Dataframe 
3. Creating a Subset of Dataframe 
4. Creating For-Loops 
5. Exporting Dataframe


# Importing Libraries 

import pandas as pd
import numpy as np 
import os 

# Importing Dataframe 

path = r'C:\Users\emily\OneDrive\Documents\Career Foundry 2022\Data Immersion\Achievment 4\Instacart Basket Analysis'

df_ords_prods_merged = pd.read_pickle(os.path.join(path, 'Data', 'Prepared Data', 'orders_products_merged.pkl'))

# Exercise 4.7 

# Creating a Subset 

#Subset of dataframe 

df = df_ords_prods_merged[:1000000]

df.shape

df.head()

#Define a function

def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

#Apply the function 

df['price_range'] = df.apply(price_label, axis=1)

df['price_range'].value_counts(dropna = False)

df['prices'].max()

#Testing the same function on entire dataset with .loc

df_ords_prods_merged.loc[df_ords_prods_merged['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_ords_prods_merged.loc[(df_ords_prods_merged['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df_ords_prods_merged.loc[df_ords_prods_merged['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_ords_prods_merged['price_range_loc'].value_counts(dropna =False)

df_ords_prods_merged['orders_day_of_week'].value_counts(dropna = False)

# Creating a For-Loop 

#Running a For-Loop 

result = []

for value in df_ords_prods_merged["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

result

#Combining For-Loop with orders_products_merged df 

df_ords_prods_merged['busiest_day'] = result

#Counting column frequency for busiest_day

df_ords_prods_merged['busiest_day'].value_counts(dropna = False)

# Task 4.7 Questions 

1. If you haven’t done so already, complete the instructions in the Exercise for creating the “price_label” and “busiest_day” columns.

2. Suppose your clients have changed their minds about the labels you created in your “busiest_day” column. Now, they want “Busiest day” to become “Busiest days” (plural). This label should correspond with the two busiest days of the week as opposed to the single busiest day. At the same time, they’d also like to know the two slowest days. Create a new column for this using a suitable method.

#Running a For-Loop for Top 2 busiest days 

result = []

for value in df_ords_prods_merged["orders_day_of_week"]:
  if value == 0 or value == 1:
    result.append("Busiest days")
  elif value == 3 or value == 4:
    result.append("Least busiest days")
  else:
    result.append("Regularly busy")

result

3. Check the values of this new column for accuracy. Note any observations in markdown format.

Answer: Regularly busy count has decreased in numbers. 

df_ords_prods_merged['busiest_days'] = result

df_ords_prods_merged['busiest_days'].value_counts(dropna = False)

4. When too many users make Instacart orders at the same time, the app freezes. The senior technical officer at Instacart wants you to identify the busiest hours of the day. Rather than by hour, they want periods of time labeled “Most orders,” “Average orders,” and “Fewest orders.” Create a new column containing these labels called “busiest_period_of_day.”

df_ords_prods_merged.head()

#Changing order_hour_of_day from floating to Integer 

df_ords_prods_merged['order_hour_of_day'] = df_ords_prods_merged['order_hour_of_day'].astype('int')

df_ords_prods_merged['order_hour_of_day'].dtype

#Checking column frequency of order_hour_of_day 

df_ords_prods_merged['order_hour_of_day'].value_counts(dropna = False)

#Running a For-Loop for Top busiest hours of the day 

result = []

for value in df_ords_prods_merged["order_hour_of_day"]:
  if value >= 9 and value <= 16:
    result.append("Most orders")
  elif value < 6:
    result.append("Fewest orders")
  else:
    result.append("Average orders")

result

df_ords_prods_merged.columns

5. Print the frequency for this new column.

#Combining For-Loop with orders_products_merged df

df_ords_prods_merged['Most_orders'] = result

df_ords_prods_merged['Most_orders'].value_counts(dropna = False)

# Exporting Dataframe 

# Import to PKL

df_ords_prods_merged.to_pickle(os.path.join(path, 'Data','Prepared Data', 'orders_products_merged_new.pkl'))

df_ords_prods_merged.shape
