==> Scripts/8. Creating Visualizations Part 1.ipynb <==
# Task 4.9 Intro to Data Visualizations With Python 

# Contents 

1. Importing Libraries 
2. Importing Customer data frame 
3. Data Wrangling 
4. Data Consistency Checks 
5. Merging Data frames 

# Task 4.6 Part 1 Questions 

1. Download the customer data set and add it to your “Original Data” folder.
2. Create a new notebook in your “Scripts” folder for part 1 of this task.
3. Import your analysis libraries, as well as your new customer data set as a dataframe.

# Importing Libraries 

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

# Importing customers.csv data frame 

path = r'C:\Users\emily\OneDrive\Documents\Career Foundry 2022\Data Immersion\Achievment 4\Instacart Basket Analysis'

cust = pd.read_csv(os.path.join(path, 'Data', 'Original Data', 'customers.csv'))

# Data Wrangling 

4. Wrangle the data so that it follows consistent logic; for example, rename columns with illogical names and drop columns that don’t add anything to your analysis.

cust.head()

cust.shape


#Renaming surnam column to be consistent with first name column

cust.rename(columns = {'Surnam':'Last Name'}, inplace = True)

#Renaming STATE column to State

cust.rename(columns = {'STATE':'State'}, inplace = True)

#Renaming n_dependants column to be more explanatory and correcting spelling error

cust.rename(columns = {'n_dependants':'no_of_dependents'}, inplace = True)

#Renaming Gender column for consistency

cust.rename(columns = {'Gender':'gender'}, inplace = True)

#Renaming State column for consistency

cust.rename(columns = {'State':'state'}, inplace = True)

#Renaming Age column for consistency

cust.rename(columns = {'Age':'age'}, inplace = True)

#Renaming fam_status column for clearer understanding

cust.rename(columns = {'fam_status':'family_status'}, inplace = True)

cust.head()

# Data Consistency Checks 

5. Complete the fundamental data quality and consistency checks you’ve learned throughout this Achievement; for example, check for and address missing values and duplicates, and convert any mixed-type data.

cust.describe()

#Checking Data Types 

cust.dtypes

#Changing 'user_id' from integer to string type 

cust['user_id'] = cust['user_id'].astype('str')

#Checking for missing values (NaN)

cust.isnull().sum()

cust_nan=cust[cust['First Name'].isnull()==True]

cust_nan

It is best to keep the 11,259 NaN values in the First_name columns as is since each customer has a user_id value assigned to them. This means we can still link the data frames together. 

#Checking for Mixed- data types 

for col in cust.columns.tolist():
  weird = (cust[[col]].applymap(type) != cust[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (cust[weird]) > 0:
    print (col)

First Name column contains mixed-data types. 

#Converting First Name column from Object to String 
cust['First Name'] = cust['First Name'].astype('str')

#Checking for Duplicates in Data frame 

cust[cust.duplicated()]

No duplicates have been found in the customer.csv data frame. 

#Reducing data types to save RAM before merging the data frames together

cust['age'] = cust['age'].astype('int8')
cust['no_of_dependents'] = cust['no_of_dependents'].astype('int8')
cust['income'] = cust['income'].astype('int32')

# Merging Data frames 

6. Combine your customer data with the rest of your prepared Instacart data. Tip: Make sure the key columns are of the same data type! Hint: Make sure the key columns are the same data type! 

#Importing orders_products_merged_complete.pkl

ords_prods = pd.read_pickle(os.path.join(path, 'Data','Prepared Data','orders_products_merged_complete.pkl'))

ords_prods.head()


#Checking the size of current dataframe 

ords_prods.shape

#Checking data type for key column = user_id

ords_prods.dtypes

#Changing user_id column from integer to string

ords_prods['user_id'] = ords_prods['user_id'].astype('str')

Reducing space for more RAM

ords_prods['product_id'] = ords_prods['product_id'].astype('int16')

ords_prods['order_number'] = ords_prods['order_number'].astype('int8')

ords_prods['orders_day_of_week'] = ords_prods['orders_day_of_week'].astype('int8')

ords_prods['order_hour_of_day'] = ords_prods['order_hour_of_day'].astype('int8')

ords_prods['add_to_cart_order'] = ords_prods['add_to_cart_order'].astype('int8')

ords_prods['reordered'] = ords_prods['reordered'].astype('int8')

ords_prods['aisle_id'] = ords_prods['aisle_id'].astype('int16')

ords_prods['department_id'] = ords_prods['department_id'].astype('int8')

ords_prods['prices'] = ords_prods['prices'].astype('float16')

ords_prods['days_since_last_order_placed'] = ords_prods['days_since_last_order_placed'].astype('float16')

ords_prods['avg_price'] = ords_prods['avg_price'].astype('float16')

ords_prods['median_order_frequency'] = ords_prods['median_order_frequency'].astype('float16')

#Verifying user_id is now a string 
ords_prods.dtypes

#Merging customer.csv dataframe into orders_products_merge_complete.pkl

ords_prods_cust = ords_prods.merge(cust, on ='user_id')

ords_prods_cust.head()

ords_prods_cust.shape

Ensuring that 9 new columns were added to the merge. There is a total of 10 columns in the original customer.csv but they are merged via the user_id column ( total 9 columns merged). Merge is proven successful. 

7. Ensure your notebook contains logical titles, section headings, and descriptive code comments.
8. Export this new dataframe as a pickle file so you can continue to use it in the second part of this task.

#Exporting Dataframe 

ords_prods_cust.to_pickle(os.path.join(path,'Data', 'Prepared Data', 'complete_instacart_data.pkl'))
